+++ src/main/java/mission1/step3/car/Car.java
@@ -1,0 +1,36 @@
+package mission1.step3.car;
+
+import mission1.step3.command.Command;
+
+import java.util.List;
+
+import static mission1.step3.command.Command.GO;
+
+public class Car {
+
+    private List<Command> commandList;
+
+    private Car(){};
+
+    public static Car createCar() {
+        return new Car();
+    }
+
+    public void setCommandList(List<Command> commandList) {
+        this.commandList = commandList;
+    }
+
+    public int getCarProgress(int currentLocation) {
+        int currentProgress = 0;
+
+        for (int i = 0; i <= currentLocation; i++) {
+            if (this.commandList.get(i).equals(GO)) {
+                currentProgress++;
+            }
+        }
+
+        return currentProgress;
+    }
+}
+
+
