+++ src/main/java/mission1/step3/command/Command.java
@@ -1,0 +1,5 @@
+package mission1.step3.command;
+
+public enum Command {
+    GO, STOP;
+}
