+++ src/main/java/mission1/step4/Main.java
@@ -1,0 +1,10 @@
+package mission1.step4;
+
+import mission1.step4.controller.CarRaceController;
+
+public class Main {
+    public static void main(String[] args) {
+        CarRaceController.run();
+    }
+}
+
