+++ src/main/java/mission1/step4/algorithm/MoveStrategy.java
@@ -1,0 +1,7 @@
+package mission1.step4.algorithm;
+
+public interface MoveStrategy {
+    public boolean movable(int randomValue);
+}
+
+
