--- src/main/java/step3/command/Command.java
@@ -1,5 +1,0 @@
-package step3.command;
-
-public enum Command {
-    GO, STOP;
-}
